Move all user related routes to their own file, link in using blueprints
